Don't require embedding unimplemented server
See: https://github.com/grpc/grpc-go/issues/3669

Opting into the default is sensible for servers that are often embedded in other servers -- for instance the gRPC health server.

Given that the Rode server is currently intended to run as a standalone service, and not planned to be embedded or available as a library, this default
doesn't apply. It also prevents us from embedding structs with partial implementations of the larger server interface -- the unimplemented server
already contains all of the exported functions, leading to ambiguity. We could solve this by having the rode server delegate to the other implementation, but
this adds some boilerplate that isn't desirable.